+++ web/packages/pyxis-app/src/components/organisms/ShowLog/index.ts
@@ -1,2 +1,3 @@
+export * from './PostShowLogEditorModal';
 export * from './PostShowLogEntryCard';
 export * from './PostShowLogPanel';